requirements.txt
# データ処理
pandas==2.2.2
openpyxl==3.1.2
beautifulsoup4==4.12.2

# ユーティリティ
requests==2.31.0
python-dotenv==1.0.1
ics==0.7.2

# テスト
pytest==7.4.3
pytest-cov==4.1.0
